Change PackingSizeType to size_t to conform to Image class

+++ source/llassetgen/include/llassetgen/Packing.h
@@ -11,7 +11,7 @@
 #include <llassetgen/llassetgen_api.h>
 
 namespace llassetgen {
-    using PackingSizeType = unsigned int;
+    using PackingSizeType = size_t;
 
     /*
      * Describes the packing of a texture atlas.
